Refactor: Move MarkdownDialog and dependencies to ui/dialogs
- Moved `MarkdownDialog.kt` from the `utilities` package to a new `ui/dialogs` package to improve project structure.
- Moved the `CustomClickableSpan.kt` utility from the `utilities` package to the `ui/dialogs` package to resolve a build failure in the 'lite' flavor and co-locate it with its dependent component.
- Updated package declarations in the moved files.
- Updated all import statements in `ChallengeHelper.kt` and `CourseStepFragment.kt` to reflect the new locations.
- Removed a redundant import from `MarkdownDialog.kt` after the move.
